requirements.txt
beautifulsoup4
FlagEmbedding==1.3.5
langchain-core
langchain-groq
langchain-text-splitters
playwright
python-dotenv
qdrant-client
streamlit
torch
transformers>=4.46,<5
tokenizers<0.22
tqdm
